+++ lib/kuby/docker/assets_phase.rb
@@ -1,6 +1,6 @@
 module Kuby
   module Docker
-    class AssetsPhase < Phase
+    class AssetsPhase < Layer
       def apply_to(dockerfile)
         dockerfile.run(
           'bundle', 'exec', 'rake', 'assets:precompile'