+++ lib/kuby/docker/layer.rb
@@ -1,17 +1,18 @@
 module Kuby
   module Docker
-    class Phase
+    class Layer
       attr_reader :definition
 
       def initialize(definition)
         @definition = definition
       end
 
-      private
-
-      def app
-        definition.app
+      def apply_to(dockerfile)
+        raise NotImplementedError,
+          "#{__method__} must be defined in derived classes"
       end
+
+      private
 
       def metadata
         definition.docker.metadata