+++ lib/kuby/kubernetes/plugins/rails_app/mysql.rb
@@ -118,10 +118,6 @@ module Kuby
             definition.kubernetes
           end
 
-          def app
-            definition.app
-          end
-
           private
 
           def config