+++ lib/kuby/railtie.rb
@@ -7,10 +7,6 @@ module Kuby
       load File.expand_path(File.join('tasks', 'kuby.rake'), __dir__)
     end
 
-    initializer 'kuby.startup' do |_app|
-      Kuby.logger = Kuby::BasicLogger.new(STDERR)
-    end
-
     initializer 'kuby.health_check_middleware' do |app|
       app.middleware.use Kuby::Middleware::HealthCheck
     end